Formation system basically works
The party's positions are tracked with a 2D array in the BattleController, and the new formation actions are special action nodes that work on the same framework as the others. The mechanics of the action are described in the "execute" function of the nodes, which the BattleController can pass its position array to in order to generate a new one with the new formations. The BattleController can use the coordinates from the new array to tell the party Sprites where to put themselves, and the party character nodes also track the row that they're in. Action nodes now have a row_limit property, which the CombatMenu can use to make sure only usable actions are available.