Worked on Action Options
Added three selectable options to the punch node and changed some of the framework of how the action nodes interact with the CharacterNode to display these options. Right now it only works for the Punch, the other two samples still aren't really at all functional. Also added the old BattleScene's attack stuff to the BattleController and finally changed Stamina and Mana to just AP.