Various tweaks, you can win now
Added more basic functionality to the enemies, a rough game over state, and remade the PartyInventory. Enemies can now attack by randomly selecting a character in the front row and punching them. Whenever a character dies, the BattleController checks through the character's team and if everyone is dead it clears the scene. The PartyInventory is a child of the BattleController representing a shared inventory that all party characters can access through a pointer. It has a function to create ActionLists based on its current contents that can be roughly slotted into the existing action node system. Right now the party starts with four stamina potions, which is just enough to carry the basic dummy fight to the end.